Fix: Re-ducksパターンの修正
