--- langdemy-app/src/components/pages/SignIn/actions.js
@@ -1,12 +1,0 @@
-export const SIGN_IN = "SIGN_IN";
-export const signInAction = (userState) => {
-  return {
-    type: "SIGN_IN",
-    payload: {
-      isSignedIn: true,
-      role: userState.role,
-      uid: userState.uid,
-      userName: userState.uid
-    },
-  };
-};
