--- langdemy-app/src/components/pages/SignUp/actions.js
@@ -1,7 +1,0 @@
-export const SIGN_UP = "SIGN_UP";
-export const signUpAction = (userState) => {
-  return {
-    type: "SIGN_UP",
-    payload: userState,
-  };
-};
