+++ langdemy-app/src/reducks/store/rootReducer.js
@@ -1,0 +1,6 @@
+import { usersReducer as users } from "../users/reducers"
+
+
+export default {
+  users,
+}
