+++ langdemy-app/src/reducks/users/reducers.js
@@ -1,10 +1,11 @@
 // 会員登録の情報を保持するモデルの宣言
 // actionをインポート
 import * as Actions from "./actions";
-import initialState from "../../../store/initialState";
+import initialState from "../store/initialState";
 
-export const signUpReducer = (state = initialState.users, action) => {
+export const usersReducer = (state = initialState.users, action) => {
   switch (action.type) {
+    case Actions.SIGN_IN:
     case Actions.SIGN_UP:
       return {
         ...state,
