--- langdemy-app/src/store/rootReducer.js
@@ -1,8 +1,0 @@
-import { signUpReducer } from  "../components/pages/SignUp/reducers";
-import { signInReducer } from "../components/pages/SignIn/reducers"
-
-
-export default {
-  signUpReducer,
-  signInReducer
-}
